Big Nujel update
Added the Nujel standalone interpreter as a target to the normale Makefiles, fixing a lot of minor issues.
Work also began on a simple stdlib for nujel written in nujel as well as support for docstrings to make the interpreter kinda self documenting, although those are still quite early in development.
BSD License got removed from the Nujel Interpreter since it is now quite intermingled with the rest of the codebase, making a separate release unlikely for now.