Only render visible chunk faces if the GL version allows for that
Specifically, if running on desktop GL or GLES with the
GL_EXT_multi_draw_arrays extension, the glMultiDrawArrays function is
available and used to draw one or more of the 6 cube face directions
within a chunk with a single call.
This requires the vertex buffer be made of 6 individual sub-buffers, so
chunk vertex list construction is now done per-side, to be uploaded
consecutively into the GPU-side buffer.

If glMultiDrawArrays is not available, gracefully falls back to just
glDrawArrays'ing the (now differently organized) buffer.
Technically you could draw each of the required sides using a
glDrawArrays each, but this is very costly especially on WebGL where GL
calls are expensive.