Added chunk fadeIn animation
This was made necessary due to the last commit which made the amount of chunks
being generated each frame timing dependent, which meant that slow machines
like the X220/RPI4 would only generate the minimum of 4 chunks per frame, which
led to many chunks just popping in. This fadeIn animation hides the popping in
making everything look (IMO) a lot better.