Fixed WASM chunk Rendering!
Sadly this meant some changes to the blockRenderer. The problem was
with indexing an uniform array by an attribute, which is not universally supported.
So the easiest fix was to use multiple draw calls and rebind a single uniform before
every call. Maube there is a more efficient way, but right now it escapes me. This also
meant the removal of the multiDraw code.